Shifting the ToyJob to the `_tests` module (since it is generic)
