+++ tests/archiving/test_import.py
@@ -4,22 +4,8 @@ from pyiron_base import Project
 from pyiron_base.archiving.import_archive import getdir, extract_archive
 from pandas._testing import assert_frame_equal
 from filecmp import dircmp
-from pyiron_base import PythonTemplateJob
 from shutil import rmtree
-from pyiron_base._tests import PyironTestCase
-
-
-class ToyJob(PythonTemplateJob):
-    def __init__(self, project, job_name):
-        """A toyjob to test export/import functionalities."""
-        super(ToyJob, self).__init__(project, job_name)
-        self.input['input_energy'] = 100
-
-    # This function is executed
-    def run_static(self):
-        with self.project_hdf5.open("output/generic") as h5out:
-            h5out["energy_tot"] = self.input["input_energy"]
-        self.status.finished = True
+from pyiron_base._tests import PyironTestCase, ToyJob
 
 
 class TestUnpacking(PyironTestCase):
